Use %lld for e_using_number_as_bool_nr
Add casts where emitting this error since varnumber_T isn't always a
long long.  This fixes the below test failures on armel/armhf/mipsel:

    From test_vim9_expr.vim:
    Found errors in Test_expr3_fails():
    command line..script /<<PKGBUILDDIR>>/src/vim-gtk3/testdir/runtest.vim[468]..function RunTheTest[39]..Test_expr3_fails[8]..CheckDefExecFailure line 7: ['assert_equal(false, Record(1) && Record(4) && Record(0))']: Expected 'E1023: Using a Number as a Bool: 4' but got 'E1023: Using a Number as a Bool: 0': ['assert_equal(false, Record(1) && Record(4) && Record(0))']

+++ src/typval.c
@@ -180,7 +180,7 @@ tv_get_bool_or_number_chk(typval_T *varp, int *denote, int want_bool)
 	    if (in_vim9script() && want_bool && varp->vval.v_number != 0
 						   && varp->vval.v_number != 1)
 	    {
-		semsg(_(e_using_number_as_bool_nr), varp->vval.v_number);
+		semsg(_(e_using_number_as_bool_nr), (long long)varp->vval.v_number);
 		break;
 	    }
 	    return varp->vval.v_number;
